+++ libs/components/lookup/src/app/public/modules/country-field/fixtures/country-field-no-form.component.fixture.ts
@@ -28,6 +28,8 @@ export class CountryFieldNoFormTestComponent {
 
   public isRequired: boolean = false;
 
+  public supportedCountryISOs: string[];
+
   public countryChanged(country: SkyCountryFieldCountry): void {
     return;
   }