+++ libs/components/lookup/src/app/public/modules/country-field/fixtures/country-field-reactive.component.fixture.ts
@@ -37,6 +37,8 @@ export class CountryFieldReactiveTestComponent implements OnInit {
 
   public initiallizeToUndefined: boolean = false;
 
+  public supportedCountryISOs: string[];
+
   public set isDisabled(value: boolean) {
     this._isDisabled = value;