+++ libs/components/lookup/src/app/public/modules/country-field/fixtures/country-field.component.fixture.ts
@@ -4,6 +4,10 @@ import {
 } from '@angular/core';
 
 import {
+  NgModel
+} from '@angular/forms';
+
+import {
   SkyCountryFieldComponent
 } from '../country-field.component';
 
@@ -22,6 +26,9 @@ export class CountryFieldTestComponent {
   })
   public countryFieldComponent: SkyCountryFieldComponent;
 
+  @ViewChild(NgModel)
+  public ngModel: NgModel;
+
   public defaultCountry: string;
 
   public isDisabled: boolean = false;
@@ -29,6 +36,8 @@ export class CountryFieldTestComponent {
   public isRequired: boolean = false;
 
   public modelValue: SkyCountryFieldCountry;
+
+  public supportedCountryISOs: string[];
 
   public countryChanged(country: SkyCountryFieldCountry): void {
     return;